Moved errors.py outside of api
