+++ dds_web/api/project.py
@@ -24,7 +24,7 @@ from dds_web import auth, db
 from dds_web.database import models
 from dds_web.api.api_s3_connector import ApiS3Connector
 from dds_web.api.db_connector import DBConnector
-from dds_web.api.errors import (
+from dds_web.errors import (
     DDSArgumentError,
     DatabaseError,
     AccessDeniedError,
