+++ dds_web/api/s3.py
@@ -13,7 +13,7 @@ import flask
 # Own modules
 from dds_web import auth
 from dds_web.api.api_s3_connector import ApiS3Connector
-from dds_web.api.errors import (
+from dds_web.errors import (
     S3ProjectNotFoundError,
 )
 from dds_web.api.schemas import project_schemas
