+++ dds_web/api/schemas/project_schemas.py
@@ -14,7 +14,7 @@ import marshmallow
 import sqlalchemy
 
 # Own modules
-from dds_web.api import errors as ddserr
+from dds_web import errors as ddserr
 from dds_web import auth, db
 from dds_web.database import models
 from dds_web.api import api_s3_connector
