+++ dds_web/api/schemas/user_schemas.py
@@ -13,7 +13,7 @@ import sqlalchemy
 
 # Own modules
 from dds_web import db
-from dds_web.api import errors as ddserr
+from dds_web import errors as ddserr
 from dds_web import auth
 import dds_web.security.auth
 from dds_web.database import models
