+++ tests/__init__.py
@@ -3,7 +3,7 @@
 from base64 import b64encode
 from urllib.parse import quote_plus
 import json
-import dds_web.api.errors as ddserr
+import dds_web.errors as ddserr
 import dds_web.database
 import flask
 import flask_login
