+++ tests/test_user_delete.py
@@ -12,7 +12,7 @@ from dds_web import db
 from dds_web.database import models
 
 import dds_web.utils
-import dds_web.api.errors as ddserr
+import dds_web.errors as ddserr
 import tests
